Add silence module to silence output from VPR
Since many processes may be running at once, the set of output streams from the
VPR processes are getting mixed together and is not of much use.  By surrounding
code with a 'with Silence():' construct, all output from the contained code is
silenced.

Add timing of placement generation and combination.
Fixes #1.